Struct servo_css_parser::types::MediaList

pub struct MediaList {
    pub media_queries: Vec<MediaQuery>,
}

A type that encapsulates a media query list.

Fields
media_queries: Vec<MediaQuery>

The list of media queries.

Methods

impl MediaList

fn empty() -> MediaList

Create an empty MediaList.

impl MediaList

fn evaluate(&self, device: &Device, quirks_mode: QuirksMode) -> bool

Evaluate a whole MediaList against Device.

fn is_empty(&self) -> bool

Whether this MediaList contains no media queries.

fn append_medium(&mut self, context: &ParserContext, new_medium: &str) -> bool

Append a new media query item to the media list. https://drafts.csswg.org/cssom/#dom-medialist-appendmedium

Returns true if added, false if fail to parse the medium string.

fn delete_medium(&mut self, context: &ParserContext, old_medium: &str) -> bool

Delete a media query from the media list. https://drafts.csswg.org/cssom/#dom-medialist-deletemedium

Returns true if found and deleted, false otherwise.