Struct servo_css_parser::types::ServoStylesheet

pub struct ServoStylesheet {
    pub contents: StylesheetContents,
    pub shared_lock: SharedRwLock,
    pub media: Arc<Locked<MediaList>>,
    pub disabled: AtomicBool,
}

The structure servo uses to represent a stylesheet.

Fields
contents: StylesheetContents

The contents of this stylesheet.

shared_lock: SharedRwLock

The lock used for objects inside this stylesheet

media: Arc<Locked<MediaList>>

List of media associated with the Stylesheet.

disabled: AtomicBool

Whether this stylesheet should be disabled.

Methods

impl Stylesheet

fn from_bytes<R>(
bytes: &[u8],
url_data: ServoUrl,
protocol_encoding_label: Option<&str>,
environment_encoding: Option<&'static Encoding>,
origin: Origin,
media: MediaList,
shared_lock: SharedRwLock,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: &R,
quirks_mode: QuirksMode
) -> Stylesheet where
R: ParseErrorReporter,

Parse a stylesheet from a set of bytes, potentially received over the network.

Takes care of decoding the network bytes and forwards the resulting string to Stylesheet::from_str.

fn update_from_bytes<R>(
existing: &Stylesheet,
bytes: &[u8],
protocol_encoding_label: Option<&str>,
environment_encoding: Option<&'static Encoding>,
url_data: ServoUrl,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: &R
) where
R: ParseErrorReporter,

Updates an empty stylesheet with a set of bytes that reached over the network.

impl Stylesheet

fn update_from_str<R>(
existing: &Stylesheet,
css: &str,
url_data: ServoUrl,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: &R,
line_number_offset: u32
) where
R: ParseErrorReporter,

Updates an empty stylesheet from a given string of text.

fn from_str<R>(
css: &str,
url_data: ServoUrl,
origin: Origin,
media: Arc<Locked<MediaList>>,
shared_lock: SharedRwLock,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: &R,
quirks_mode: QuirksMode,
line_number_offset: u32
) -> Stylesheet where
R: ParseErrorReporter,

Creates an empty stylesheet and parses it with a given base url, origin and media.

Effectively creates a new stylesheet and forwards the hard work to Stylesheet::update_from_str.

fn disabled(&self) -> bool

Returns whether the stylesheet has been explicitly disabled through the CSSOM.

fn set_disabled(&self, disabled: bool) -> bool

Records that the stylesheet has been explicitly disabled through the CSSOM.

Returns whether the the call resulted in a change in disabled state.

Disabled stylesheets remain in the document, but their rules are not added to the Stylist.